+++ src/graphology/src/data.js
@@ -29,8 +29,6 @@ MixedNodeData.prototype.clear = function () {
   this.inDegree = 0;
   this.outDegree = 0;
   this.undirectedDegree = 0;
-  this.directedSelfLoops = 0;
-  this.undirectedSelfLoops = 0;
 
   // Indices
   this.in = {};
@@ -57,7 +55,6 @@ DirectedNodeData.prototype.clear = function () {
   // Degrees
   this.inDegree = 0;
   this.outDegree = 0;
-  this.directedSelfLoops = 0;
 
   // Indices
   this.in = {};
@@ -82,7 +79,6 @@ export function UndirectedNodeData(key, attributes) {
 UndirectedNodeData.prototype.clear = function () {
   // Degrees
   this.undirectedDegree = 0;
-  this.undirectedSelfLoops = 0;
 
   // Indices
   this.undirected = {};